Graph: add Multi-Topology support
Graph supports IPv4 or IPv6 topology but with IS-IS Multi-Topology
it is possible to advertise both IPv4 and IPv6 topologies. Recent
versions of BGP-LS speaker also support this Multi-Topology.
This patch changes the Graph yang model (including the revision date)
to handle IPv4 and IPv6 simultaneously on Edges and Vertices for
IS-IS Multi-Topology.

JIRA: BGPCEP-1010

Signed-off-by: Olivier Dugeon <olivier.dugeon@orange.com>
Change-Id: Icfcf9e8a27f5f9d3fd9a928dba8d3adffaf8de50