Split up BasePCEPSessionProposalFactory
We are concentrating what amounts to a utility method into a strict
hierarchy, conflating a few arguments. At the end of the day we want to
disconnect PCEPTimerProposal from the factory, just as we do not want
to mention capabilities.

This takes a step towards that goal, eventually leading to users to
provide PCEPSessionProposalFactory on their own.

Change-Id: I1f31a800d961daa04cf2f46292cbc338d07106b5
JIRA: BGCEP-962
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>