WIP: Remove BundleContext from pcep-tunnel-provider
The wiring between topology provider and tunnel provider is broken, as
the topology provider does not register InstructionScheduler into OSGi
SR.

Rather than bringing in that dependency, define a PCEPTopologyExtender
contract, which is injected by PCEPTopologyTracker as a set of optional
dependencies.

This provides the explicit lifecycle hook, allowing tunnel provider to
react to topologies being started and stopped.

Change-Id: I1553b1916f2f1f52e9cf61058df55300b297e0e6
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>