Improve RouteDistinguisherUtil
Do not muck with NormalizedNodes.findNode(), as we have the same
functionlity directly available.

Eliminate RDType -- it is a noble idea, but all we really need is just
valid constants. This improves both paths, as the arguments are now
known compile-time constants.

Finally improve nullness annotations, eliminating the need for a
requireNonNull() in flowspec.

Change-Id: I176db96b783649de166797ee84bf919e68a57196
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>