Clean up RibSupport registration
The RIBSupport integration shows a number of cases of accidental
duplicity with the evolution over the years. At the end of the day, each
RIBSupport knows to what AFI/SAFI it is bound -- on both Binding and DOM
layers.

Use that information to simplify interactions, improving ergonomics
across the board, as well as ditching a source of Binding->DOM identity
translation.

JIRA: BGPCEP-1016
Change-Id: I19377f25cf6adb1a9ddc170913961f948c10938c
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>