PCEP: Correct Exception when router send an Empty RRO
When a non-delegated LSP is reported by a router to ODL-PCEP, if the tunnel is
Down, the RRO contains no valid path i.e. RRO is empty but present. When this
RRO is parsed by PCEPReportedRouteObjectParser.java, it raised the non null non
empty verification which raised an exception. Unfortunately, this exception
forces ODL-PCEP to close the session without a particular reason. Then, the router
re-establishes the PCEP session and send the same PcReport message with the
empty RRO. This cause the same exception, thus there is an infite loop of Open,
Report Close PCEP messages which stops only when the LSP becomes UP or Active,
thus reporting a non empty RRO. The problem occurs a startup i.e. the LSP is
Down before ODL-PCEP started or during PCEP session i.e. LSP goes Down.

This patch corrects this infinite loop by parsing an empty RRO without raising
an exception.

JIRA: BGPCEP-1032

Change-Id: I5acfc32f41c6213fc85d80085f3ac8666b9b779f
Signed-off-by: Olivier Dugeon <olivier.dugeon@orange.com>