Remove deprecated InstanceIdentifier
Instance Identifier has been mark as deprecated since release 14.0.0.
This patch update the code to remove deprecated functions.

To achieve the objective, a replacement of the `firstKeyOf()' method
was added to the DefaultInstanceReference class in the concepts bundle.
This method is a ccopy/paste from the same method of the deprecated
Instance Identifier class.

Signed-off-by: Olivier Dugeon <olivier.dugeon@orange.com>
Change-Id: Ia358760849167a5f3791281c9df6172ba7c8c14d