Use import-by-revision in openconfig-api
OC is naughty, as it renamed bgp-well-known-std-community to
BGP_WELL_KNOWN_STD_COMMUNITY, hence we cannot survive with an upgraded
model. This shows up in integration tests with tcpe.

To fix this, we'll just use import-by-revision in openconfig-api, so
that we disregard any other revisions.

Change-Id: I9be98c0d9e69439e525c450f57f47aca43e9b6e2
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>
(cherry picked from commit be8f59f605d9229e041218990942858093466798)