PCEP: Update Path Binding to RFC9604
Update ietf-stateful yang model and Path Binding parser and
serializer according to RFC9605. Main modifications are:
- TLV code point is now 55 (IANA allocation)
- Binding Type is encoded other 8 bits instead of 16 bits
- Flags fields (8 bits) has been introduced
- Srv6 parser / serializer (Binding Type = 2)

A new Binding Type enumaration has been added to the yang
model to ease Path Binding management.

Note that BT=3 a.k.a. SRv6 with Endpoint Behavior will be
introduced latter with the support of SRv6 in PCEP.

Signed-off-by: Olivier Dugeon <olivier.dugeon@orange.com>
Change-Id: I9cb753d7f4724dc1331c1db4db734fbcd768f97c