Algo: Correct Path computation with Include Route
If an Include Router (i.e. a list of IP address) is specify in
path constraints, the path computation is not handle correctly
the additive metrics constraints. In fact, during merging phase,
additive metrics of the various part of the computed path are not
summed and original constraints are not adjusted accordingly.

E.g. a constraint metric of 20 with 2 IP adresses for Include
Route will try to compute 3 parts with constraint metric 20 for
each and a final comuted metric equal to the first one, instead:
- compute the part from source to first IP addres with metric 20
- compute the part form first IP to second IP address with metric
  20 - the metric budget of the first part
- compute the last part from 2nd IP address to destination with
  metric 20 - the metric budget of the two first parts
- return the path from source to destination with computed metric
  equal to the computed metric of the 3 parts.

Change-Id: I2b060ad2017205e0de48d43ffba2e719b1baff94
Signed-off-by: Olivier Dugeon <olivier.dugeon@orange.com>