Refactor PathManagerListener
This is a thorough modernization, taking advantage of Java 21 as well as
the underlying DataObject* semantics:

We are listening only on PcepNodeConfig, which eliminates the need to filter
through Node changes: the infrastructure does that for us.

For each change, we recover the corresponding NodeKey via
DataTreeModification.path().getFirstKeyOf(Node.class).

We then use an enhanced switch to dispatch on DataObjectModification
type. This would nominally have three cases (Deleted,Modified,Written), but
our logic works on dataAfter(), so this gets reduced to just two cases.

We also use DataObjectModification.getModifiedChildren(Class) to get the
child modifications without having to do the filtering/collecting
ourselves. Here it is critical we do not check for emptiness, but just
interate over the returned collection: if it is empty we will do nothing
anyway.

The final bit here is ConfiguredLspKey recovery in the deleted case:
this is stored in DataObjectModification.step(), as we statically know
what the change is: so we use coerceKeyStep() and get the key without
ever looking at dataBefore().

Change-Id: Ia4c323c92b5fef71daef00b7330f73423e642cef
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>